+++ src/Config/PuliFileLocator.php
@@ -19,7 +19,7 @@ use Puli\Repository\ResourceRepository;
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class PuliFileLocator implements ChainableFileLocatorInterface
+class PuliFileLocator implements ChainableFileLocator
 {
     /**
      * @var ResourceRepository