+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+* 1.0.0-alpha2 (@release_date@)
+
+ * Renamed `ChainableFileLocator` to `ChainableFileLocatorImpl`
+ * Renamed `ChainableFileLocatorInterface` to `ChainableFileLocator`
+ 
 * 1.0.0-alpha1 (2014-12-03)
 
  * first alpha release